+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 9.0.0
+- Add support for `Swift Package Manager`. [#127](https://github.com/BendingSpoons/tempura-swift/pull/127)
+- Align `Tempura` and `TempuraTesting` versions. [#127](https://github.com/BendingSpoons/tempura-swift/pull/127)
+
 ## Tempura 7.1.0
 
 - Allow to init Navigation with custom `RoutableProvider`. [#125](https://github.com/BendingSpoons/tempura-swift/pull/125)